[refactor] #25 - 세션 목록 화면 뷰 분리
